requirements.txt
fastapi
uvicorn
joblib
scikit-surprise
pandas
boto3
requests
